invoice-creation-service/src/invoices/invoices.controller.ts
import { Controller, Get, Post, Body, Param, Query } from '@nestjs/common';
import { InvoicesService } from './invoices.service';
import { CreateInvoiceDto } from './dto/create-invoice.dto';
import { Invoice } from './schemas/invoice.schema';

@Controller('invoices')
export class InvoicesController {
  constructor(private readonly invoicesService: InvoicesService) {}

  @Post()
  async create(@Body() createInvoiceDto: CreateInvoiceDto): Promise<Invoice> {
    return this.invoicesService.create(createInvoiceDto);
  }

  @Get()
  async findAll(@Query() query: any): Promise<Invoice[]> {
    const filters = {};
    if (query.startDate && query.endDate) {
      filters['date'] = {
        $gte: new Date(query.startDate),
        $lte: new Date(query.endDate),
      };
    }
    return this.invoicesService.findAll(filters);
  }

  @Get(':id')
  async findOne(@Param('id') id: string): Promise<Invoice> {
    return this.invoicesService.findOne(id);
  }
}
